updated worker log
