+++ redash/authentication/org_resolving.py
@@ -2,9 +2,9 @@ import logging
 
 from flask import g, request
 from werkzeug.local import LocalProxy
-
 from redash.models import Organization
 from redash.utils.format import CustomFormatter
+from redash.utils.job_filter import CurrentJobFilter
 def _get_current_org():
     if "org" in g:
         return g.org
@@ -17,6 +17,7 @@ def _get_current_org():
     g.org = Organization.get_by_slug(slug)
     logging.debug("Current organization: %s (slug: %s)", g.org, slug)
     CustomFormatter.org_id = g.org
+    CurrentJobFilter.org_id = g.org
     return g.org
